fix: update permissions and refactor package structure for consistency
