--- app/src/main/java/hcmute/edu/vn/linhvalocvabao/selfalarmproject/adapters/MusicAdapter.java
@@ -1,83 +1,0 @@
-package hcmute.edu.vn.linhvalocvabao.selfalarmproject.adapters;
-
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.bumptech.glide.Glide;
-
-import java.util.List;
-
-import hcmute.edu.vn.linhvalocvabao.selfalarmproject.R;
-import hcmute.edu.vn.linhvalocvabao.selfalarmproject.data.model.Music;
-
-public class MusicAdapter extends RecyclerView.Adapter<MusicAdapter.MusicViewHolder> {
-
-    private List<Music> musicList;
-    private Context context;
-
-    public MusicAdapter(List<Music> musicList, Context context) {
-        this.musicList = musicList;
-        this.context = context;
-    }
-
-    @NonNull
-    @Override
-    public MusicViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.item_music, parent, false);
-        return new MusicViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull MusicViewHolder holder, int position) {
-        Music music = musicList.get(position);
-        
-        holder.titleTextView.setText(music.getTitle());
-        holder.artistTextView.setText(music.getArtists());
-        
-        // Load thumbnail image
-        if (music.getThumbnailM() != null && !music.getThumbnailM().isEmpty()) {
-            Glide.with(context)
-                    .load(music.getThumbnailM())
-                    .placeholder(R.drawable.ic_launcher_background)
-                    .error(R.drawable.ic_launcher_background)
-                    .into(holder.thumbnailImageView);
-        }
-        
-        // Set click listener
-        holder.itemView.setOnClickListener(v -> {
-            Toast.makeText(context, "Selected: " + music.getTitle(), Toast.LENGTH_SHORT).show();
-        });
-    }
-
-    @Override
-    public int getItemCount() {
-        return musicList.size();
-    }
-
-    public void updateData(List<Music> newMusicList) {
-        this.musicList = newMusicList;
-        notifyDataSetChanged();
-    }
-
-    static class MusicViewHolder extends RecyclerView.ViewHolder {
-        ImageView thumbnailImageView;
-        TextView titleTextView;
-        TextView artistTextView;
-
-        public MusicViewHolder(@NonNull View itemView) {
-            super(itemView);
-            thumbnailImageView = itemView.findViewById(R.id.thumbnailImageView);
-            titleTextView = itemView.findViewById(R.id.titleTextView);
-            artistTextView = itemView.findViewById(R.id.artistTextView);
-        }
-    }
-}
